+++ src/api/user.js
@@ -5,6 +5,7 @@ const { hashString } = require('../utils/hash');
 class Person {
   constructor(username) {
     this.username = username;
+    this.data = {};
   }
 
   create(firstName, familyName, email, details, password) {
@@ -15,6 +16,7 @@ class Person {
       email,
       details,
       password,
+      roles: ['editor', 'commenter'],
     };
 
     return new Promise((resolve, reject) => user.findOne(
@@ -26,6 +28,7 @@ class Person {
         return user.create(personObj, (createErr) => {
           if (createErr) return reject(createErr);
 
+          this.data = personObj;
           return resolve(personObj);
         });
       },
@@ -38,6 +41,7 @@ class Person {
         if (err) return reject(err);
         if (doc === null) return reject(new Error('User not found'));
 
+        this.data = doc;
         return resolve(new Person(doc.username));
       });
     });
@@ -49,6 +53,7 @@ class Person {
         if (err) return reject(err);
         if (doc === null) return reject(new Error('User not found!'));
 
+        this.data = doc;
         return resolve(doc);
       });
     });