+++ src/utils/db.js
@@ -3,12 +3,12 @@ const { createHash } = require('crypto');
 
 const {
   DB_USER,
-  DB_PWD,
   DB_HOST,
   DB_PORT,
   DB_NAME,
 } = process.env;
 
+const DB_PWD = encodeURIComponent(process.env.DB_PWD);
 let uri = `mongodb+srv://${DB_USER}:${DB_PWD}@${DB_HOST}/${DB_NAME}?retryWrites=true&w=majority`;
 if (process.env.STANDARD_CONNECTION) {
   uri = `mongodb://${DB_USER}:${DB_PWD}@${DB_HOST}:${DB_PORT}/${DB_NAME}?authSource=admin`;
@@ -19,6 +19,11 @@ const news = mongoose.model('new', mongoose.Schema({
   title: String,
   desc: String,
   path: String,
+  pullRequest: [{
+    _id: mongoose.Types.ObjectId,
+    personId: String,
+    diff: String,
+  }],
   author: mongoose.Types.ObjectId,
   date: Date,
   dateLastmod: Date,
@@ -31,6 +36,20 @@ const news = mongoose.model('new', mongoose.Schema({
     isBasedOn: [String],
   },
   content: String,
+  editors: [String],
+  comments: [{
+    _id: mongoose.Types.ObjectId,
+    personId: String,
+    content: String,
+    comments: [{
+      _id: mongoose.Types.ObjectId,
+      personId: String,
+      content: String,
+      reactions: [[String, Number]],
+    }],
+    reactions: [[String, Number]],
+  }],
+  reactions: [[String, Number]],
 }, { versionKey: false }));
 
 const userSchema = mongoose.Schema({
@@ -46,12 +65,14 @@ const userSchema = mongoose.Schema({
     gender: Number,
   },
   password: String,
+  roles: [String],
 }, { versionKey: false });
 
 userSchema.post('findOne', (person) => {
   if (person !== null) delete person._doc.password;
 });
 
+// eslint-disable-next-line func-names
 userSchema.pre('save', function (next) {
   if (this.password && this.isModified('password')) {
     this.password = createHash('sha256').update(this.password).digest('hex');
@@ -62,11 +83,7 @@ userSchema.pre('save', function (next) {
 
 const user = mongoose.model('user', userSchema);
 
-mongoose.connect(uri, {
-  useNewUrlParser: true,
-  useUnifiedTopology: true,
-  useFindAndModify: false,
-});
+mongoose.connect(uri);
 
 mongoose.connection.on('open', () => console.log('Database connected succesfully'));